Don't parse clock expressions that are commented out
Presumably there are also other org expressions that this code will still find
clocks in that are not really there.  For example (note leading whitespace):

 # CLOCK: [2019-06-04 Tue 16:00]--[2019-06-04 Tue 17:00] =>  1:00

The above is not commented (because this form of comment must start at the first
character on a line), so this fix still leaves that treated as a clock
expression.  I imagine org does not treat it as such, so that's probably a bad
thing -- but it's unchanged behaviour and I'm only trying here to fix the case
of explicitly commented code.